refactor(reporting): aligns file structure with level 1 namespace
